Refactored game networking a bit
