+++ src/client/gamenet.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void GameNetStartThread();
